// src/components/transactions/types.ts
interface IOrganizer {
  id: string;
  username: string;
  first_name: string;
  last_name: string;
}

interface ITransaction {
  id: string;
}

interface IVoucher {
  id: string;
  code: string;
  discount_amount: number;
  start_date?: string;
  end_date?: string;
  max_usage?: number;
  current_usage?: number;
  event_id?: string;
}

interface IReview {
  id: string;
}

export interface IEventDetails {
  id: string;
  name: string;
  start_date: string;
  end_date: string;
  description: string;
  event_image: string;
  location: string;
  price: number;
  total_seats: number;
  remaining_seats: number;
  category: string;
  organizer_id: string;
  organizer: IOrganizer;
  transactions: ITransaction[];
  voucher: IVoucher[];
  review: IReview[];
  created_at?: string;
  updated_at?: string;
}

export interface TransactionFormValues {
  event_id: string;
  quantity: number;
  attend_date: string;
  use_voucher: boolean;
  voucher_code: string;
  use_coupon: boolean;
  coupon_code: string;
  use_points: boolean;
  points_to_use: number;
  payment_method: string;
}